pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "markforge"
version = "0.1.0"
description = "Text watermark/stamp tool (CLI-first, GUI-ready)."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "Grant" }]
keywords = ["watermark", "image", "pillow", "typer", "cli"]
classifiers = [
  "License :: Other/Proprietary License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Utilities",
]
dependencies = [
  "pillow>=10.0.0",
  "typer>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/remgrandt/markforge"
Repository = "https://github.com/remgrandt/markforge"

[project.scripts]
markforge = "markforge.cli:app"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.6.0",
]
gui = [
  # Choose later if/when you add a real GUI:
  # "pyside6>=6.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/markforge"]
include = ["src/markforge/static/**"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
